Empty parent class to start building functions into
